Перевёл расчёт стоимости доставки на BigDecimal

+++ commerce/delivery/src/main/java/ru/yandex/practicum/controller/DeliveryController.java
@@ -23,7 +23,7 @@ public class DeliveryController implements DeliveryClient {
 
     @Override
     public BigDecimal deliveryCost(OrderDto order) {
-        return BigDecimal.valueOf(deliveryService.deliveryCost(order));
+        return deliveryService.deliveryCost(order);
     }
 
     @Override
